250411 작업 내용
- [x]  보스 패턴 완성
- [x]  보스 UI 추가
- [x]  타일 오브젝트 충돌 기능 수정
- [x]  방 3개 오브젝트 배치 완료
- [x]  패시브 아이템 1종 추가

+++ Isaac/Isaac/CUIBar.cpp
@@ -82,7 +82,7 @@ void CUIBar::Render(HDC hDC)
 	{
 		
 		BitBlt(hDC,
-			600 + 40 * i,
+			600 + 40 * (int)i,
 			30,
 			40, 40,
 			hMemDC,
